refactor: improve test folder format and move csrf helper function to setup test file
