+++ tests/initialisation/app.test.js
@@ -1,17 +1,6 @@
-require("../testSetup");
 const request = require("supertest");
 const app = require("../../app");
-
-async function getCsrfToken(route = "/create") {
-  const response = await request(app).get(route);
-  const csrfTokenMatch = response.text.match(/name="_csrf" value="([^"]+)"/);
-  if (!csrfTokenMatch) {
-    throw new Error(`Failed to extract CSRF token from route: ${route}`);
-  }
-  const csrfToken = csrfTokenMatch[1];
-  const cookie = response.headers["set-cookie"];
-  return { csrfToken, cookie };
-}
+const { getCsrfToken } = require("../setup/testSetup");
 
 describe("App Initialisation", () => {
   test("should use URL-encoded middleware and return 404 for unknown POST route", async () => {
